Refactoring: created new 'utils.py' file with general useful methods
